pyproject.toml
[tool.poetry]
name = "mihama"
version = "0.7.0"
description = "osv.dev API clone for on-premise usage"
authors = ["Manabu Niseki <manabu.niseki@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
aiometer = "^0.5"
arq = "^0.25"
elasticsearch = { extras = ["async"], version = "^8.13" }
elasticsearch-dsl = { extras = ["async"], version = "^8.13" }
fastapi = "^0.110"
gitpython = "^3.1"
gunicorn = "^21.2"
httpx = "^0.27.0"
loguru = "^0.7"
orjson = "^3.10"
packageurl-python = "^0.15"
pydantic = "^2.6"
pyhumps = "^3.8"
redis = { extras = ["hiredis"], version = "^5.0" }
returns = { extras = ["compatible-mypy"], version = "^0.22" }
semver = "^3.0"
typer = "^0.12"
uvicorn = { extras = ["standard"], version = "^0.29" }

[tool.poetry.group.dev.dependencies]
ci-py = "^1.0"
nest-asyncio = "^1.6"
pre-commit = "^3.7"
pytest = "^8.1"
pytest-asyncio = "^0.23"
pytest-cov = "^5.0"
pytest-docker = "^3.1"
pytest-mock = "^3.14"
pytest-randomly = "^3.15"
types-requests = "2.31"
vcrpy = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.poetry.scripts]
mihama = "mihama.cli:app"

[tool.mypy]
plugins = ["pydantic.mypy", "returns.contrib.mypy.returns_plugin"]
ignore_missing_imports = true

[tool.ruff.lint]
select = [
    "ASYNC", # flake8-async
    "B",     # flake8-bugbear
    "C",     # flake8-comprehensions
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "PT",    # flake8-pytest-style
    "RET",   # flake8-return
    "RUF",   # Ruff-specific rules
    "SIM",   # flake8-simplify
    "T20",   # flake8-print
    "UP007", # non-pep604-annotation
    "W",     # pycodestyle warnings
]
ignore = ["E501"]
